// valentines-invite/src/types.ts
export interface DateOption {
    id: string;
    title: string;
    image: string;
    locations: string[];
    foodOptions?: string[];
    activities?: string[];
    hasOwnFood?: boolean;
    customInput?: boolean;
}

export interface DateSelection {
    category: string;
    location: string;
    foodOption?: string;
    activities: string[];
    bringOwnFood?: boolean;
    comments: string;
    customImage?: string;
}